refactor: 기존의 PathFinder 클래스를 추상화
JGraphT에 의존하는 PathFinder를 추상화
