+++ README.md
@@ -17,6 +17,7 @@
 
 [여기](./subway.http)에서 API 요청 형식을 확인할 수 있습니다.
 
+
 ## 구현 기능 목록
 ### 노선에 역 등록
 - [x] 노선을 최초 등록할 때 두 역을 동시에 등록한다.
@@ -29,12 +30,14 @@
 - [x] 노선에 갈래길이 존재할 수 없다.
 - [x] 역 간 거리는 언제나 앙의 정수여야 한다.
 
+
 ### 노선에 역 제거
 - [x] 노선에서 역을 제거하는 경우 나머지 역의 연결이 재배치된다.
 - [x] 노선의 마지막 두 역 중 하나를 제거하는 경우 노선도 삭제된다.
 
 
 - [x] 노선에 등록되지 않은 역을 삭제할 수 없다.
+
 
 ### 경로 조회
 - [x] 경로 조회 시 요금 정보를 포함하여 응답한다.
@@ -43,6 +46,7 @@
 
 - [x] 노선에 아직 연결되지 않은 역과의 경로를 조회할 수 없다.
 - [x] 노선끼리 연결되지 않은 상황에서 다른 노선의 두 역 간의 경로를 조회할 수 없다.
+
 
 ## 프로그래밍 요구 사항
 - 프로덕션 환경은 [로컬 h2](/src/test/resources/application.properties)에서 동작한다. 별도의 h2가 설치 및 실행된 상태에서 프로그램이 동작한다.
