+++ src/main/java/subway/application/PathService.java
@@ -9,9 +9,10 @@ import subway.dao.SectionDao;
 import subway.dao.StationDao;
 import subway.domain.fare.FareCalculator;
 import subway.domain.path.PathException;
-import subway.domain.path.PathFinder;
+import subway.domain.path.JGraphTPathFinder;
 import subway.domain.Sections;
 import subway.domain.Station;
+import subway.domain.path.PathFinder;
 import subway.ui.dto.PathRequest;
 import subway.ui.dto.PathResponse;
 
@@ -35,7 +36,7 @@ public class PathService {
         Station target = stationDao.findById(pathRequest.getTargetStationId())
             .orElseThrow(() -> new PathException("등록되지 않은 역과의 경로는 찾을 수 없습니다."));
 
-        PathFinder pathFinder = new PathFinder(new Sections(sectionDao.findAll()), source, target);
+        PathFinder pathFinder = new JGraphTPathFinder(new Sections(sectionDao.findAll()), source, target);
         List<Station> vertices = pathFinder.getPathVerticies();
         Sections edges = pathFinder.getPathEdges();
